refactor(studio): extract cutover-eligibility checks to unblock filesize cap
